completed functionality of profile page
