+++ package.json
@@ -25,6 +25,7 @@
     "react-rangeslider": "^2.2.0",
     "react-redux": "^5.0.7",
     "react-responsive-carousel": "^3.1.41",
+    "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.5",
     "react-slick": "^0.23.1",
